+++ gui/omni/assets/StartMenu.qml
@@ -50,6 +50,7 @@ Rectangle {
                 Repeater {
                     model: [
                         {name: "Filez Explorer", icon: "📁", id: "explorer"},
+                        {name: "OmniClerk", icon: "💼", id: "omniclerk"},
                         {name: "OmniSec", icon: "☠️", id: "omnisec"},
                         {name: "OmniAudio", icon: "🎧", id: "omniaudio"},
                         {name: "OmniVision", icon: "🎬", id: "omnivision"},